pyproject.toml
[project]
name = "alto-terminal"
version = "0.1.0"
description = "LiveKit audio client for streaming audio without the agents framework"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "livekit>=0.17.0",
    "livekit-api>=1.1.0",
    "numpy>=1.20.0",
    "sounddevice>=0.4.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]
